update tests for mpc
