+++ tests/test_dof.py
@@ -130,120 +130,6 @@ def test_loadcase():
         assert "top" in sh[0]
 
 
-def test_mpc():
-
-    mesh = fe.Cube(n=3)
-    element = fe.Hexahedron()
-    quadrature = fe.quadrature.GaussLegendre(order=1, dim=3)
-
-    mesh.points = np.vstack((mesh.points, [2, 0, 0]))
-    mesh.update(mesh.cells)
-
-    region = fe.Region(mesh, element, quadrature)
-
-    u = fe.FieldContainer([fe.Field(region, dim=3)])
-    F = u.extract()
-
-    umat = fe.constitution.NeoHooke(mu=1.0, bulk=2.0)
-
-    f0 = lambda x: np.isclose(x, 0)
-    f1 = lambda x: np.isclose(x, 1)
-    f2 = lambda x: np.isclose(x, 2)
-
-    boundaries = {}
-    boundaries["left"] = fe.Boundary(u[0], fx=f0)
-    boundaries["right"] = fe.Boundary(u[0], fx=f2, skip=(1, 0, 0))
-    boundaries["move"] = fe.Boundary(u[0], fx=f2, skip=(0, 1, 1), value=0.5)
-
-    mpc = fe.Boundary(u[0], fx=f1).points
-    cpoint = mesh.npoints - 1
-
-    RBE2 = fe.MultiPointConstraint(mesh, points=mpc, centerpoint=cpoint)
-    K_RBE2 = RBE2.stiffness()
-    r_RBE2 = RBE2.residuals(u)
-
-    linearform = fe.IntegralForm(umat.gradient(F), u, region.dV)
-    r = linearform.assemble() + r_RBE2
-
-    bilinearform = fe.IntegralForm(umat.hessian(F), u, region.dV, u)
-    K = bilinearform.assemble() + K_RBE2
-
-    assert r.shape == (84, 1)
-    assert K.shape == (84, 84)
-
-
-def pre_mpc_mixed(point, values):
-
-    mesh = fe.mesh.Cube(n=3)
-    element = fe.element.Hexahedron()
-    quadrature = fe.quadrature.GaussLegendre(order=1, dim=3)
-
-    mesh.points = np.vstack((mesh.points, point))
-    mesh.update(mesh.cells)
-
-    region = fe.Region(mesh, element, quadrature)
-    dV = region.dV
-
-    mesh0 = fe.mesh.convert(mesh, order=0)
-    element0 = fe.element.ConstantHexahedron()
-    region0 = fe.Region(mesh0, element0, quadrature, grad=False)
-
-    displacement = fe.Field(region, dim=3)
-    pressure = fe.Field(region0)
-    volumeratio = fe.Field(region0, values=1)
-
-    displacement.values[-1] = values
-
-    fields = fe.FieldContainer([displacement, pressure, volumeratio])
-
-    F, p, J = fields.extract()
-
-    nh = fe.NeoHooke(mu=1.0, bulk=2.0)
-    umat = fe.ThreeFieldVariation(nh)
-
-    f0 = lambda x: np.isclose(x, 0)
-    f1 = lambda x: np.isclose(x, 1)
-    f2 = lambda x: np.isclose(x, 2)
-
-    boundaries = {}
-    boundaries["left"] = fe.Boundary(displacement, fx=f0)
-    boundaries["right"] = fe.Boundary(displacement, fx=f2, skip=(1, 0, 0))
-    boundaries["move"] = fe.Boundary(displacement, fx=f2, skip=(0, 1, 1), value=0.5)
-
-    mpc = fe.Boundary(displacement, fx=f1).points
-    cpoint = mesh.npoints - 1
-
-    RBE2 = fe.MultiPointConstraint(mesh, points=mpc, centerpoint=cpoint)
-    CONT = fe.MultiPointContact(mesh, points=mpc, centerpoint=cpoint)
-    K_RBE2 = RBE2.stiffness()
-    r_RBE2 = RBE2.residuals(fields)
-
-    K_CONT = CONT.stiffness(fields)
-    r_CONT = CONT.residuals(fields)
-
-    assert K_RBE2.shape == K_CONT.shape
-    assert r_RBE2.shape == r_CONT.shape
-
-    linearform = fe.IntegralForm(umat.gradient([F, p, J]), fields, dV)
-    r = linearform.assemble()
-
-    r_RBE2.resize(*r.shape)
-    r = r + r_RBE2
-
-    bilinearform = fe.IntegralForm(umat.hessian([F, p, J]), fields, dV, fields)
-    K = bilinearform.assemble()
-
-    K_RBE2.resize(*K.shape)
-    K = K + K_RBE2
-
-
-def test_mpc_mixed():
-    pre_mpc_mixed(point=[2, 0, 0], values=[0, 0, 0])
-    pre_mpc_mixed(point=[2, 0, 0], values=[-5, 0, 0])
-
-
 if __name__ == "__main__":
     test_boundary()
     test_loadcase()
-    test_mpc()
-    test_mpc_mixed()
